archive/core-components/aletheia-backend/requirements.txt
# =====================
# 核心框架
# =====================
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.6.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# =====================
# 数据库
# =====================
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis
redis[hiredis]==5.0.1

# TimescaleDB
timescaledb==0.0.4

# 向量数据库
qdrant-client==1.7.0

# =====================
# AI/ML框架
# =====================
# LLM
langchain==0.1.3
langchain-openai==0.0.3
openai==1.9.0

# ML for Fake News Detection
scikit-learn>=1.3.0
joblib>=1.3.0

# OCR (暂时禁用,Python 3.12不兼容)
# paddleocr==2.7.0.3
# paddlepaddle==2.6.0
# easyocr==1.7.1 (暂时禁用,依赖torch)

# ASR
# openai-whisper==20231117 (暂时禁用,依赖torch)

# 视觉
opencv-python==4.9.0.80
pillow==10.2.0
# mediapipe==0.10.32 (暂时禁用,依赖torch)

# NLP
jieba==0.42.1
numpy==1.26.3
scipy==1.12.0

# =====================
# 消息队列
# =====================
kafka-python==2.0.2
celery[redis]==5.3.6

# =====================
# 爬虫
# =====================
httpx==0.26.0
beautifulsoup4==4.12.3
playwright==1.41.1
fake-useragent==1.4.0
brotli==1.1.0

# =====================
# 证据库
# =====================
trafilatura>=1.6.0
lxml>=4.9.0
opensearch-py>=2.4.0
readability-lxml>=0.8.1

# =====================
# 认证与安全
# =====================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
email-validator==2.1.0

# =====================
# 工具库
# =====================
python-dotenv==1.0.1
pytz==2024.1
pydantic-extra-types==2.3.0
typing-extensions==4.9.0
pyyaml==6.0.1

# =====================
# 监控与日志
# =====================
sentry-sdk[fastapi]==1.40.0
prometheus-client==0.19.0
loguru==0.7.2

# =====================
# 测试
# =====================
pytest==8.4.2
pytest-asyncio==0.24.0
pytest-cov==4.1.0
hypothesis==6.98.3
httpx==0.26.0

# =====================
# 开发工具
# =====================
black==24.1.1
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0
